add logout alert

+++ src/app/admin/header-admin/header-admin.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { AuthService } from '../../users/auth.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-header-admin',
@@ -13,7 +14,24 @@ export class HeaderAdminComponent {
   private _authService = inject(AuthService);
   private _router = inject(Router)
   onLogOut(){
-      this._authService.purgeAuth();
-      this._router.navigate(['/login']);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You want to logout?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, Sure!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Success",
+          text: "You have been logged out!",
+          icon: "success"
+        });
+        this._authService.purgeAuth();
+        this._router.navigate(['/login']);
+      }
+    });
   }
 }
